move variable buffer into seperate file
